fixed some library issues.
all the models look nice and scale well, but the haptic rendering isn't working
on some of the models after the initial load.

can load the 5M point fireplace model and interact with it altough its a bit
slow at 100fps.

+++ pbrenderer/LYCudaHelper.cuh
@@ -17,7 +17,7 @@ namespace LYCudaHelper
 {
 	void allocateHostArray(void **devPtr, size_t size, unsigned int flag = cudaHostAllocPortable);
 	void freeHostArray(void *devPtr);
-	void allocateArray(void **devPtr, int size);
+	void allocateArray(void **devPtr, size_t size);
 	void freeArray(void *devPtr);
 	void *mapGLBufferObject(struct cudaGraphicsResource **cuda_vbo_resource);
 	void unmapGLBufferObject(struct cudaGraphicsResource *cuda_vbo_resource);
